+++ src/Common/Driver/XCache.php
@@ -8,12 +8,12 @@
 
 namespace LeanPHP\Behat\CodeCoverage\Common\Driver;
 
-use PHP_CodeCoverage_Driver as DriverInterface;
+use SebastianBergmann\CodeCoverage\Driver\Driver as DriverInterface;
 
 /**
  * XCache Driver
  *
- * {@internal Derived from PHP_CodeCoverage_Driver_Xdebug.}
+ * {@internal Derived from SebastianBergmann\CodeCoverage\Driver\Xdebug.}
  *
  * @author Anthon Pang <apang@softwaredevelopment.ca>
  */
@@ -22,25 +22,25 @@ class XCache implements DriverInterface
     /**
      * Constructor
      *
-     * @throws \PHP_CodeCoverage_Exception if PHP code coverage not enabled
+     * @throws \SebastianBergmann\CodeCoverage\RuntimeException if PHP code coverage not enabled
      */
     public function __construct()
     {
         if ( ! extension_loaded('xcache')) {
-            throw new \PHP_CodeCoverage_Exception('This driver requires XCache');
+            throw new \SebastianBergmann\CodeCoverage\RuntimeException('This driver requires XCache');
         }
 
         if (version_compare(phpversion('xcache'), '1.2.0', '<') ||
             ! ini_get('xcache.coverager')
         ) {
-            throw new \PHP_CodeCoverage_Exception('xcache.coverager=On has to be set in php.ini');
+            throw new \SebastianBergmann\CodeCoverage\RuntimeException('xcache.coverager=On has to be set in php.ini');
         }
     }
 
     /**
      * {@inheritdoc}
      */
-    public function start()
+    public function start($determineUnusedAndDead = true)
     {
         xcache_coverager_start();
     }