+++ src/Common/Report/Php.php
@@ -9,6 +9,8 @@
 namespace LeanPHP\Behat\CodeCoverage\Common\Report;
 
 use LeanPHP\Behat\CodeCoverage\Common\ReportInterface;
+use SebastianBergmann\CodeCoverage\CodeCoverage;
+use SebastianBergmann\CodeCoverage\Report\PHP as PHPReport;
 
 /**
  * PHP report
@@ -18,7 +20,7 @@ use LeanPHP\Behat\CodeCoverage\Common\ReportInterface;
 class Php implements ReportInterface
 {
     /**
-     * @var \PHP_CodeCoverage_Report_Clover
+     * @var \SebastianBergmann\CodeCoverage\Report\Clover
      */
     private $report;
 
@@ -36,14 +38,14 @@ class Php implements ReportInterface
             $options['target'] = null;
         }
 
-        $this->report = new \PHP_CodeCoverage_Report_PHP();
+        $this->report = new PHPReport();
         $this->options = $options;
     }
 
     /**
      * {@inheritdoc}
      */
-    public function process(\PHP_CodeCoverage $coverage)
+    public function process(CodeCoverage $coverage)
     {
         return $this->report->process(
             $coverage,