+++ tests/Common/Report/Crap4jTest.php
@@ -10,6 +10,7 @@ namespace LeanPHP\Behat\CodeCoverage\Common\Report;
 
 use VIPSoft\TestCase;
 use LeanPHP\Behat\CodeCoverage\Common\Report\Factory;
+use SebastianBergmann\CodeCoverage\Report\Crap4j;
 
 /**
  * Crap4j report test
@@ -20,12 +21,25 @@ class Crap4jTest extends TestCase
 {
     public function testProcess()
     {
-        if ( ! class_exists('PHP_CodeCoverage_Report_Crap4j')) {
+        if ( ! class_exists('SebastianBergmann\CodeCoverage\Report\Crap4j')) {
             $this->markTestSkipped();
 
             return;
         }
 
-        $this->markTestIncomplete();
+        $report = $this->getMockBuilder('SebastianBergmann\CodeCoverage\Report\Crap4j')
+                       ->disableOriginalConstructor()
+                       ->getMock();
+
+        $coverage = $this->createMock('SebastianBergmann\CodeCoverage\CodeCoverage');
+        $coverage->expects($this->once())
+                 ->method('getReport')
+                 ->will($this->returnValue($report));
+
+        $report = new Crap4j();
+        $result = $report->process($coverage);
+
+        $this->assertTrue(strpos($result, '<?xml version="1.0" encoding="UTF-8"?>') === 0);
+
     }
 }