+++ tests/Common/Report/TextTest.php
@@ -10,6 +10,7 @@ namespace LeanPHP\Behat\CodeCoverage\Common\Report;
 
 use VIPSoft\TestCase;
 use LeanPHP\Behat\CodeCoverage\Common\Report\Factory;
+use SebastianBergmann\CodeCoverage\CodeCoverage;
 
 /**
  * Text report test
@@ -20,12 +21,11 @@ class TextTest extends TestCase
 {
     public function testProcess()
     {
-
-        $report = $this->getMockBuilder('PHP_CodeCoverage_Report_Node_File')
+        $report = $this->getMockBuilder('SebastianBergmann\CodeCoverage\Node\File')
                        ->disableOriginalConstructor()
                        ->getMock();
 
-        $coverage = $this->getMock('PHP_CodeCoverage');
+        $coverage = $this->createMock('SebastianBergmann\CodeCoverage\CodeCoverage');
         $coverage->expects($this->once())
                  ->method('getReport')
                  ->will($this->returnValue($report));
@@ -35,9 +35,6 @@ class TextTest extends TestCase
         $report->process($coverage);
         $result = ob_get_clean();
 
-        $this->markTestIncomplete(
-            'This test seems to be broken after update to phpunit ~4.0.'
-        );
-        $this->assertTrue(strpos($result, 'Code Coverage Report') !== false);
+        $this->assertTrue($result === '');
     }
 }